pyproject.toml
[tool.poetry]
name = "python-socket-sdk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "python_socket_sdk"}]

[tool.poetry.dependencies]
python = "^3.8"
pyzmq = "^25.0.0"
protobuf = ">=3.20.0,<3.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
